And update composer.lock to match

530294df67 LocalFileTest: use encodeBlob/decodeBlob for img_metadata
4953d68866 DatabaseSqlite: fix null blobs
671d302f90 rdbms: avoid pg_escape_bytea() call-style deprecation notices
83f896f817 Improve LocalisationCache post-merge validation check
3db092b8ea Update wikimedia/remex-html to 3.0.3
ea0058a5c9 Fix the remaining Phan failures on PHP 8.1
73e84f8183 Respond to some messages from Phan on PHP 8.1
3c7b171104 Update RELEASE-NOTES-1.39
c2bccb1a1b Fix phan error when Excimer is enabled
828209c16c Add matrix: to $wgUrlProtocols
7c75c31efd build: Add phan stub for ReturnTypeWillChange
e32a9cc8c8 tests: Change assertion in AbstractSecondaryAuthenticationProviderTest
6a2596252c tests: Declare GIFMetadataExtractorTest::$mediaPath
8ca55ef7bd tests: Allow dynamic properties on PHP 8.2 for GhostFieldTestClass
45dc6cffa1 stream wrapper: Declare $context class property
613c33b223 tests: Remove assertion of spec in ParamValidatorTest::testGetTypeDef
f23c230418 libs\jsminplus: Declare JSNode::$expression
8c5996d3f0 composer.json: Updated composer/spdx-licenses from 1.5.6 to 1.5.7
4084c352ee tests: Replace dynamic property MWCallableUpdate::$_name
3e71a726d6 Upgrading cssjanus/cssjanus (v2.1.0 => v2.1.1)
a265640444 rdbms: Remove deprecation mark for $wgSharedDB
8280ca9563 installer: Split drop action out of the SQL patch for actor migration
10859035f8 tests: Handle dynamic created properties on test classes
2706b7130e SqliteMaintenance.php: Fix fatally broken instanceof check
e3f348c6c3 Localisation updates from https://translatewiki.net.
c24cc9dddb rdbms: Use DBConnRef in SelectQueryBuilder
c5e6f0d832 api/en.json: api-help-datatype-expiry add missing 'may'
57b30e4336 Localisation updates from https://translatewiki.net.
858cc51173 OutputPage: Fix undefined `['host']` in ImagePreconnect code
a0d98c9c25 Pass empty string to strlen() if schema is null for PostgresDatabase
f494353abc build: Updating npm dependencies
d5fbd15e89 Update RELEASE-NOTES-1.39
27a7479ccd SpecialRevisionDelete: Set default of '' for wpReason
b8d011fef9 Localisation updates from https://translatewiki.net.
96ca0d65f7 Fix XML dumps for content types with non-string getNativeData()
417b629417 PoolCounterRedis: Fix wrong cast, locks weren't being released
f1b5cb9034 build: Updating http-cache-semantics to 4.1.1
d42a609ab4 revisiondelete: Replace dynamic property Status::$itemStatuses
8e636165e8 Update OOUI to v0.44.5
efe421c108 Localisation updates from https://translatewiki.net.
b011fb353c skin: Restore default 'value' attribute in makeSearchButton()
15c5f0b373 ParamValidator: Improve paramvalidator-help-multi-max message
af6b6ec1d8 Clear the statsd data buffer regardless of StatsdServer config
85d559f00b Localisation updates from https://translatewiki.net.
8217542bc5 WikiImporter: do not fail if upload entry in dump lacks 'text' tag
41f5389d45 Localisation updates from https://translatewiki.net.
eec149a9f8 Update RELEASE-NOTES-1.39
580e1d300b UnregisteredLocalFile: Don't call MimeAnalyzer if no path
bac194b619 TempFSFile: Use a WeakMap for reference tracking if available
b4420b3ba6 LinksUpdate: Use DB key for category links table
eb8ba34695 i18n: Add no to fallback chain of nb and nn
5799d47bcc Prep 1.39.2
dabdc50801 Start 1.39.3
20c96a9b2a GlobalFunctions: Remove check for MEDIAWIKI constant
577fbce3d4 API: Fix query+allimages user parameter description
6bc70d2f2c SpecialEditTags: Set default of '' for wpReason

Depends-On: If0554322bb41a70f85ed142da921d31ccaeb0bea

+++ includes/specials/SpecialWatchlist.php
@@ -823,7 +823,7 @@ class SpecialWatchlist extends ChangesListSpecialPage {
 		] ) );
 		asort( $hours );
 
-		$select = new XmlSelect( 'days', 'days', (float)( $selectedHours / 24 ) );
+		$select = new XmlSelect( 'days', 'days', $selectedHours / 24 );
 
 		foreach ( $hours as $value ) {
 			if ( $value < 24 ) {
